refactor: user controller
