+++ src/repositories/evaluation-template-content-repository.ts
@@ -1,0 +1,21 @@
+import prisma from "../utils/prisma"
+import { type Prisma } from "@prisma/client"
+
+export const getById = async (id: number) => {
+  return await prisma.evaluation_template_contents.findUnique({
+    where: {
+      id,
+    },
+  })
+}
+
+export const getAllByFilters = async (where: Prisma.evaluation_template_contentsWhereInput) => {
+  return await prisma.evaluation_template_contents.findMany({
+    select: {
+      id: true,
+      name: true,
+      description: true,
+    },
+    where,
+  })
+}
